Move shared components from common to base, cleanup imports
