+++ packages/base/src/Checkbox/CheckboxList/AppCheckboxList.tsx
@@ -12,7 +12,7 @@ import {
 import { StyledFormGroup, StyledViewMore } from "./AppCheckboxList.styles";
 import { AppSelectAllCheckbox } from "../components/AppSelectAllCheckbox";
 import { StyledLabel } from "../AppCheckbox.styles";
-import { StyledLabelBox } from "@benbeck764/react-components-common";
+import StyledLabelBox from "../../_common/LabelBox";
 
 export type AppCheckboxListProps = {
   items: CheckboxItem[];
