+++ packages/base/src/_common/EllipsingText.tsx
