+++ packages/base/src/index.ts
@@ -2,6 +2,13 @@
 export * from "./_theme/overrides.theme";
 export { default as CustomThemeProvider } from "./_theme/CustomThemeProvider";
 
+// Common
+export { default as CloseButton } from "./_common/CloseButton";
+export { default as StyledEllipsingTextContainer } from "./_common/EllipsingText";
+export { default as StyledLabelBox } from "./_common/LabelBox";
+export { default as TypographySkeleton } from "./_common/TypographySkeleton";
+
+// Base Components
 export { default as AppBreadcrumbs } from "./Breadcrumbs/AppBreadcrumbs";
 export * from "./Breadcrumbs/AppBreadcrumbs.props";
 export { default as AppButton } from "./Button/AppButton";
