+++ packages/common/src/index.ts
@@ -1,7 +1,2 @@
 export * from "./hooks";
 export * from "./utilities";
-
-export { default as CloseButton } from "./components/CloseButton";
-export { default as StyledEllipsingTextContainer } from "./components/EllipsingText";
-export { default as StyledLabelBox } from "./components/LabelBox";
-export { default as TypographySkeleton } from "./components/TypographySkeleton";
