+++ packages/dropzone/src/FileDropzone/AppFileDropzone.styles.ts
@@ -1,5 +1,5 @@
 import { styled, Box, Stack } from "@mui/material";
-import { CloseButton } from "@benbeck764/react-components-common";
+import { CloseButton } from "@benbeck764/react-components";
 
 export const StyledDropzoneWrapper = styled(Stack)(({ theme }) => ({
   padding: theme.spacing(2),
